Refactored plotting software
Split plot module into plot.monthly and plot.summary modules
