+++ main.py
@@ -1,3 +1,8 @@
-import user_io
 import statements
-import plot
+import plot.monthly
+import plot.summary
+
+
+print('Generating plots...')
+plot.monthly.do_it(statements.data)
+plot.summary.do_it(statements.data)
